Remove broken data_pos mechanism
Trying to pack additional data into a completed packet can't be done
without a race condition. The `port_step` at the end of the function
will call `bridge_start_in`, handing control of the buffer to the
bridge. From that point on, the bridge is free to schedule a DMA
transfer from a high-priority interrupt, and it is undefined whether any
modification to the buffer will be seen by the DMA controller
transferring to the SoC. And you have no way of knowing what the DMA
controller read, so it will lose data.

The port code may not write to the buffer or call bridge_start_in again
until the completion callback is called.