fixes error thrown when both wifi-iface disabled
There was an error being thrown when trying to scan for available
networks while both wireless wifi-iface configs are disabled. The
solution to this was to check the state of the first wifi-iface.disabled
and enable it if it was not already.

However, trying to scan immediately after enabling the wifi-iface still
threw the error OR would return with empty results, depending on a
delay. The fix _this_ issue, the scanWifi function would recursively
defer resolving until producing results or after at least 13 attempts
(which would generally indicate there are actually no available
networks). The number of attempts is arbirtrary but produces consistent results.