Make Windows auto-install WinUSB driver.
Apply the Windows Compatible ID descriptor to the CDC interfaces
as well as the vendor-class interface. It seems that Windows will
only automatically install the driver if *all* interface of the device
have the WCID descriptor set to WINUSB.

This will prevent the CDC serial console interface from working on
Windows.

Bumps the bcdDevice so Windows machines that have cached the descriptors
will re-load them.