Fix boot LED animation
The TCC prescaler was set to 1, not 1024, because the line meant to set
CTRLA.ENABLE also cleared the rest ot CTRLA, including the prescaler
field. However, a period of 65536 at 48MHz gives 732.4Hz, a reasonable
PWM frequency.

The TCC1 ISR failed to clear the interrupt flag, meaning that it re-
executed as soon as it exited. Thus, the `interpolate` function
generated a very high frequency sine wave as fast as it could calculate.
The obeserved frequency on the LED was due to aliasing when this wave was
sampled at the 732Hz period of the TCC, where CC0BUF is copied to CC0.
Instead, this sets the pattern period deliberately.

Division and modulo by 2^16 is much faster than by 2^16-1 because it can
be done with a bit shift and mask.