Avoid bridge state corruption when the SoC polls too fast
Increase the delay in spid and make the bridge code more resilient
against the EVSYS ISR preempting the DMA ISR when the bridge gets a new
transaction before finishing processing the last one.

There could be a better solution with more invasive changes to tell the
SoC when it is allowed to poll. This fix doesn't guarantee correctness,
and costs performance, but substantially improves stability when using
all three bridge pipes simultaneously.